[IOMC] Move public headers to interface directory
